WIP: fixed the height and width percentage annotation

+++ utils.py
@@ -1,9 +1,9 @@
 import settings
 
 
-def height_prct(percentage) -> int:
+def height_prct(percentage) -> float:
     return (settings.HEIGHT / 100) * percentage
 
 
-def width_prct(percentage) -> int:
+def width_prct(percentage) -> float:
     return (settings.WIDTH / 100) * percentage
